requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
python-dotenv
beautifulsoup4
lxml
unstructured
tiktoken
torch
transformers
huggingface-hub
pypdf